v2/client: file shuffle
