+++ v2/client/client.go
@@ -1,19 +1,10 @@
 package client
 
 import (
-	"bytes"
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/url"
-	"strings"
-
-	"github.com/docker/distribution/reference"
-
-	voucher "github.com/grafeas/voucher/v2"
 )
 
 var errNoHost = errors.New("cannot create client with empty hostname")
@@ -24,53 +15,6 @@ type Client struct {
 	httpClient *http.Client
 	username   string
 	password   string
-}
-
-// Check executes a request to a Voucher server, to the appropriate check URI, and
-// with the passed reference.Canonical. Returns a voucher.Response and an error.
-func (c *Client) Check(ctx context.Context, check string, image reference.Canonical) (voucher.Response, error) {
-	var checkResp voucher.Response
-	var buffer bytes.Buffer
-
-	err := json.NewEncoder(&buffer).Encode(voucher.Request{
-		ImageURL: image.String(),
-	})
-	if err != nil {
-		return checkResp, fmt.Errorf("could not parse image, error: %s", err)
-	}
-
-	var req *http.Request
-
-	req, err = http.NewRequestWithContext(
-		ctx,
-		http.MethodPost,
-		toVoucherCheckURL(c.hostname, check),
-		&buffer,
-	)
-	if nil != err {
-		return checkResp, err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	if c.username != "" && c.password != "" {
-		req.SetBasicAuth(c.username, c.password)
-	}
-	resp, err := c.httpClient.Do(req)
-	if nil != err {
-		return checkResp, err
-	}
-
-	if !strings.Contains(resp.Header.Get("Content-Type"), "application/json") {
-		var b []byte
-		b, err = ioutil.ReadAll(resp.Body)
-		if nil == err {
-			err = fmt.Errorf("failed to get response: %s", strings.TrimSpace(string(b)))
-		}
-		return checkResp, err
-	}
-
-	err = json.NewDecoder(resp.Body).Decode(&checkResp)
-	return checkResp, err
 }
 
 // SetBasicAuth adds the username and password to the Client struct
